database_sink/requirements.txt
psycopg2-binary
quixstreams~=2.8.0